refactor(observability): sweep routes onto domain telemetry imports
