+++ surfsense_backend/app/routes/new_chat_routes.py
@@ -52,7 +52,7 @@ from app.db import (
     get_async_session,
     shielded_async_session,
 )
-from app.observability import analytics as ph_analytics
+from app.observability.analytics import posthog as ph_analytics
 from app.schemas.new_chat import (
     AgentToolInfo,
     CancelActiveTurnResponse,
